Improve conclusion chapter by expanding some key sociological topics.
As discussed with Quim, there is a need to highlight some key
sociological topics in the conclusion chapter. I clarified some
arguments and added several paragraphs linking the socialization
of production pointed in the Introduction with the high rate of
individual turn over in the top levels of the connectivity hierarchy.

Also following Fabrizio's comments from some weeks ago I made clear
that the focus of the thesis is on how large scale cooperation works
and not about its emergence (this was not always clear in some places).